Conflicts:
	include/nat64/common/config.h
	include/nat64/mod/common/config.h
	include/nat64/mod/common/json_parser.h
	mod/common/config.c
	mod/common/nl_handler.c
	mod/stateful/Kbuild
	mod/stateful/filtering_and_updating.c
	mod/stateful/joold.c
	mod/stateful/json_parser.c
	mod/stateful/nf_hook.c
	mod/stateful/session/db.c
	mod/stateless/Kbuild
	mod/stateless/impersonator.c
	mod/stateless/json_parser.c
	mod/stateless/nf_hook.c
	usr/stateful/JsonReader.c

Still to-do:

- Review of joold, json and generic netlink.
- Refactor of global.
- Fragment database.
- Some testing.
- Update userspace app.

Eclipse counts 73 in-code TO-DOs (probably ~60 belong to this
iteration).

Not stable yet.

+++ include/nat64/usr/joold/jool_client.h
@@ -1,11 +1,14 @@
 #ifndef JOOL_CLIENT_H
 #define JOOL_CLIENT_H
 
+#include <linux/types.h>
 #include <netlink/netlink.h>
 #include <netlink/socket.h>
 #include <netlink/msg.h>
 #include <netlink/attr.h>
 #include <netlink/object.h>
+#include <netlink/genl/genl.h>
+#include <netlink/genl/ctrl.h>
 
 
 #if !defined(LIBNL_VER_NUM)
@@ -17,6 +20,6 @@
 
 	int set_updated_entries(void *data);
 	int get_updated_entries(void);
-	int jool_client_init(int (*cb)(void *, size_t size));
+	int jool_client_init(int (*cb)(void *, __u16 size));
 
 #endif
